redoing everything to support themes
reworking it so that the theming comes from the app tag rather than the
body tag. this broke v-dropdown so i've moved that project into this one
to support applying the tags to the root vue element.

tweaked some of the light theme colors too while I was at it